requirements_onnx.txt
streamlit
onnxruntime
opencv-python
pillow
numpy
matplotlib